+++ drizzle/0038_marvelous_jackpot.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "bookings" ADD COLUMN "booking_location" jsonb;--> statement-breakpoint
+ALTER TABLE "event_types" ADD COLUMN "locations" jsonb DEFAULT '[]'::jsonb NOT NULL;